### 2024-08-01 #7065 @editorialbot
[PRE REVIEW]: Annotate-Lab: Simplifying Image Annotation
**Submitting author:** @sumn2u (<a href="http://orcid.org/0000-0002-4345-1050">Suman Kunwar</a>)
**Repository:** https://github.com/sumn2u/annotate-lab
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.0.0
