### 2024-08-01 956c27b
fix configuration image  reference
