### 2024-08-04 cdf8887
fix: server/Dockerfile to reduce vulnerabilities
The following vulnerabilities are fixed with an upgrade:
- https://snyk.io/vuln/SNYK-DEBIAN12-ZLIB-6008963
- https://snyk.io/vuln/SNYK-DEBIAN12-GCC12-5901316
- https://snyk.io/vuln/SNYK-DEBIAN12-GLIBC-1546991
- https://snyk.io/vuln/SNYK-DEBIAN12-UTILLINUX-2401083
- https://snyk.io/vuln/SNYK-DEBIAN12-UTILLINUX-2401083
