+++ server/Dockerfile
@@ -1,4 +1,4 @@
-FROM python:3.9-slim
+FROM python:3.13.0rc1-slim
 
 WORKDIR /app
 
