### 2024-09-10 #7065 @boisgera
Peter is the author of [EcoAssist](https://joss.theoj.org/papers/10.21105/joss.05581); his review will likely start in ~1 month.

Thank you Peter! 🙏
