### 2024-09-10 #7210 @editorialbot
[REVIEW]: Annotate-Lab: Simplifying Image Annotation
**Submitting author:** @sumn2u (<a href="http://orcid.org/0000-0002-4345-1050">Suman Kunwar</a>)
**Repository:** https://github.com/sumn2u/annotate-lab
**Branch with paper.md** (empty if default branch): paper
**Version:** 2.0.3
