### 2024-10-09 #7210 @PetervanLunteren
Hi @sumn2u,

I’ve taken a close look at your tool, and I want to commend you on the impressive work and extensive documentation. I have gone through the reviewer’s checklist and the [Review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), and I noticed a few minor points that I would appreciate some clarification on:

1. **Annotation project options:** At the start of an annotation project, users can choose between “Image classification” and “Image segmentation.” I’d like to point out that “Image classification” typically refers to assigning a single label to an entire image, while the functionality of placing bounding boxes, circles, or polygons is more accurately described as “Object detection.” Would it be possible to revise this terminology to better reflect this distinction?

2. **User experience for annotating multiple images:** In scenarios where users want to annotate multiple images, particularly in bulk (potentially thousands), implementing shortcut keys could significantly enhance the user experience. During my testing, I found that the only way to move to the next image was by clicking on it in the list. Did I overlook any existing shortcuts, or is there a plan to introduce this feature in future updates?

3. **Comparison with similar tools:** In the manuscript, you mention other similar software tools such as Label Studio, VGG, COCO Annotator, Super Annotate, and CVAT. However, I found it challenging to pinpoint the main differences between your tool and these others. Is the primary distinguishing factor that your tool is open-source and community-driven? It might be helpful to elaborate on this in the manuscript to highlight what sets your tool apart.

Thank you for your time, and I look forward to your responses!
