### 2024-10-09 52cacaa
rename image classification to object detection
