### 2024-10-09 d537ce7
rename image classification to object detection
