+++ client/src/Localization/translation-en-EN.js
@@ -34,7 +34,7 @@ const translationEnEN = {
   "setup.tabs.images": "Images",
   "setup.tabs.taskinfo.task_info": "Task Information",
   "setup.tabs.taskinfo.task_choice": "Choice of Task",
-  "setup.tabs.taskinfo.task_choice_classification": "Image Classification",
+  "setup.tabs.taskinfo.task_choice_object_detection": "Object Detection",
   "setup.tabs.taskinfo.task_choice_segmentation": "Image Segmentation",
   "configuration.multiple_regions": "Can Multiple Regions Be Created?",
   "configuration.multiple_region_labels": "Multiple Region Labels Allowed?",
@@ -69,7 +69,7 @@ const translationEnEN = {
   auto_bounding_box_done: "Auto Bouding Box Selection Completed",
   comment_placeholder: "Write comment here...",
   image_tags: "Image Tags",
-  image_tags_classification_placeholder: "Image Classification",
+  image_tags_object_detection_placeholder: "Object Detection",
   loading: "Loading...",
   note: "Note",
   here: "here",
