### 2024-10-09 7cc2821
rename image classification to object detection
