### 2024-10-09 #7210 @sumn2u
Hi @PetervanLunteren , Thank you for your review and for your kind words! I truly appreciate your thoughtful feedback on both the tool and its documentation. Here's answer to your questions:

> 1. **Annotation project options:** At the start of an annotation project, users can choose between “Image classification” and “Image segmentation.” I’d like to point out that “Image classification” typically refers to assigning a single label to an entire image, while the functionality of placing bounding boxes, circles, or polygons is more accurately described as “Object detection.” Would it be possible to revise this terminology to better reflect this distinction?

It makes sense to update the terminology, as it better conveys the intended functionality. I’ve made the necessary changes and updated the [repository](https://github.com/sumn2u/annotate-lab/pull/197), [documentation](https://annotate-docs.dwaste.live/fundamentals/set-up-and-run/configuration), and manuscript accordingly. It looks like this now.

> 2. **User experience for annotating multiple images:** In scenarios where users want to annotate multiple images, particularly in bulk (potentially thousands), implementing shortcut keys could significantly enhance the user experience. During my testing, I found that the only way to move to the next image was by clicking on it in the list. Did I overlook any existing shortcuts, or is there a plan to introduce this feature in future updates?

We don’t have this feature yet, but we plan to add it in the near future.

> 3. **Comparison with similar tools:** In the manuscript, you mention other similar software tools such as Label Studio, VGG, COCO Annotator, Super Annotate, and CVAT. However, I found it challenging to pinpoint the main differences between your tool and these others. Is the primary distinguishing factor that your tool is open-source and community-driven? It might be helpful to elaborate on this in the manuscript to highlight what sets your tool apart.

Yes, our tool is indeed open-source and community-driven. Moreover, our client-server architecture enhances flexibility and scalability, which distinguishes us from other annotation tools. I have added this section in our manuscript to provide a clearer comparison and to emphasize the unique features of our tool.

Please let me know if anything is unclear.
