### 2024-10-09 346480e
fix: selection tools not being selected issue
