### 2024-10-09 8cb992c
fix: selection tools not being selected issue
