### 2024-10-19 6169bcd
fix: server/Dockerfile to reduce vulnerabilities
The following vulnerabilities are fixed with an upgrade:
- https://snyk.io/vuln/SNYK-DEBIAN12-ZLIB-6008963
- https://snyk.io/vuln/SNYK-DEBIAN12-GLIBC-1546991
- https://snyk.io/vuln/SNYK-DEBIAN12-GLIBC-1547039
- https://snyk.io/vuln/SNYK-DEBIAN12-GLIBC-1547069
- https://snyk.io/vuln/SNYK-DEBIAN12-GLIBC-1547135
