### 2024-10-23 dd64b24
added shortcuts info
