### 2024-10-23 f301c26
added shortcuts info
