### 2024-10-30 #7210 @sumn2u
Hi @boisgera, Do you happen to know when the next review will begin? I was under the impression that the entire review process would start in about seven weeks.
