### 2024-10-30 #7210 @jpcbertoldo
> Hi @boisgera, Do you happen to know when the next review will begin? I was under the impression that the entire review process would start in about seven weeks.

Hi @sumn2u, I apologize for delaying this, I had some other things to prioritize, but I will be able to deal with this in a few days. So sorry for taking long @boisgera !
