### 2024-11-13 sumn2u/annotate-lab#203 @jpcbertoldo
Re-uploading files erases the previously uploaded ones
When re-uploading files, the previous ones are erased.
My suggestion is to keep the previous uploads and add the new ones to the list.
If this is intended, it would be nice to have a warning.

ubuntu 20.04, chrome 130

I wrote this is part of my review to JOSS https://github.com/openjournals/joss-reviews/issues/7210
