### 2024-11-13 #7210 @sumn2u
Thank @jpcbertoldo  for your thoughtful review and helpful suggestions! We’re glad to hear you find annotate-lab useful. We'll prioritize working on the improvements you mentioned, like packaging the Python server, specifying Python/Node.js versions, and adding clearer SAM activation steps in the README.

We also appreciate you opening GitHub issues for the edge cases you found—it’s a big help as we refine the tool. Thanks again for your support!
