+++ paper/paper.bib
@@ -22,7 +22,7 @@
 	abstract = {Annotate better with CVAT, the industry-leading data engine for machine learning. Used and trusted by teams at any scale, for data of any scale.},
 	urldate = {2024-06-19},
 	publisher = {CVAT.ai Corporation},
-	author = {{CVAT. ai Corporation}},
+	author = {{CVAT.ai Corporation}},
 	month = jun,
 	year = {2024},
 	doi = {10.5281/ZENODO.11543564},
