### 2024-11-13 #7210 @boisgera
Hi @sumn2u,

I made a final pass on your paper draft and except for a [very small typo](https://github.com/sumn2u/annotate-lab/pull/208), I have no comment to make. 👍

Could you adress the "Additional Author Tasks" in [your post-review list](https://github.com/openjournals/joss-reviews/issues/7210#issuecomment-2474174623)?

I'll address the rest of my own tasks as soon as you give me your archive DOI and version number.

Best regards,

Sébastien
