### 2024-11-13 #7210 @sumn2u
@boisgera I checked all points. Here are the information:

- [x] Double check authors and affiliations (including ORCIDs).
    Release Version: [2.0.3](https://github.com/sumn2u/annotate-lab/releases/tag/2.0.3)
- [x] Archive the release on Zenodo/figshare/etc. and post the DOI here. [![DOI](https://zenodo.org/badge/805920466.svg)](https://doi.org/10.5281/zenodo.14147025)
- [x] Ensure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
- [x] Verify that the license listed for the archive is the same as the software license.
