### 2024-11-13 #7210 @boisgera
Hi again @sumn2u!

Could you please edit your Zenodo repo. to:

  - [ ] make the title match exactly the title of your JOSS paper ("Annotate-Lab: Simplifying Image Annotation")
  - [ ] list only yourself as the author (instead of I guess all contributors to the project; there is an other field for contributors that you can use AFAICT),
  - [ ] register your ORCID id with Zenodo (it should appear as a green badge)
  - [ ] make your Zenodo license match your GitHub project license (MIT).

If you need an example, use e.g. https://doi.org/10.5281/zenodo.13922881
