### 2024-11-13 #7210 @sumn2u
Sure @boisgera. I made all the requested changes.

- [x] make the title match exactly the title of your JOSS paper ("Annotate-Lab: Simplifying Image Annotation")
- [x] list only yourself as the author (instead of I guess all contributors to the project; there is an other field for contributors that you can use AFAICT),
- [x] register your ORCID id with Zenodo (it should appear as a green badge)
- [x]  make your Zenodo license match your GitHub project license (MIT).
